Adicionando os observers
Fazendo o parse do json e alimentando os observables  conforme chega mensagem do broker

+++ lib/components/card_plant.dart
@@ -17,7 +17,7 @@ class CardPlant extends StatefulWidget {
   final String apelidoPlanta;
   final String especiePlanta;
   final Function functionL;
-  int umidadeDoSolo;
+  double umidadeDoSolo;
   bool estadoLampada;
 
   @override
